vendo o nevoeiro mas encarando de frente

+++ admin/src/main/java/com/example/components/progress/Circle_progress.java
@@ -40,7 +40,7 @@ public class Circle_progress  extends Group{
             PseudoClass.getPseudoClass("indeterminate");
     
 
-    public Circle_progress(int radius,int size,double translate,int values){
+    public Circle_progress(int radius,int size,double translate,float values){
         ar = Arco(radius,0);
         value = new Label("0%");
         value.setPrefWidth(size);
